Open med details in full-screen modal instead of inline drawer
Match the task and appointment pattern: clicking a med row opens a
modal (or full-screen sheet on mobile) with all editable fields and
the refill workflow actions, rather than expanding an inline drawer
beneath the row.

- Extract drawer body into MedDetailModal with the same Inline*
  helpers, autosave behavior, refill workflow buttons, and
  deactivate flow.
- Slim MedRow down to a click-to-open summary row; drop the kebab
  menu (its actions all live in the modal now).
- Lift the viewing-med state to MedicationsView so a single modal
  instance hosts the detail (and the nested RefillModal).
- Update medications.spec.js to drive the modal selectors instead of
  .med-drawer.open, and to confirm the refill modal when stepping
  through the Mark refilled flow.
